.gitconfig: Add a new global "sync" target
The "sync" target is a simple play on checkout-pull-checkout-rebase.
When trying to keep in sync with a remote project, I find myself doing
this often.  Now, writing:

  git sync remote-branch

Will update the local tracking of remote-branch and rebase the
local branch you were on when you issued the command onto remote-branch.
By default, remote-branch is master.